refactor: Score 클래스 제거 후 책임 재분배
