+++ src/main/java/domain/cards/Hand.java
@@ -7,42 +7,59 @@ import java.util.List;
 public class Hand {
 
     private static final int FIRST_CARD_INDEX = 0;
+    private static final int BUST_THRESHOLD = 21;
+    private static final int DEALER_HIT_THRESHOLD = 16;
+    private static final int A_SCORE_GAP = 10;
 
     private final List<Card> cards;
-    private final Score score;
 
     public Hand() {
         this.cards = new ArrayList<>();
-        this.score = new Score();
     }
 
     public void addCard(Card card) {
         cards.add(card);
-        calculateScore();
     }
 
-    public Score calculateScore() {
-        score.sumAllCards(cards);
+    public int calculateScore() {
+        int score = sumAllCards(cards);
         if (hasAce()) {
-            score.decideScore();
+            score = decideScore(score);
         }
         return score;
+    }
+
+    private int sumAllCards(List<Card> cards) {
+        return cards.stream()
+                .mapToInt(Card::score)
+                .sum();
     }
 
     private boolean hasAce() {
         return cards.stream().anyMatch(Card::isAce);
     }
 
+    private int decideScore(int score) {
+        if (cannotBust(score + A_SCORE_GAP)) {
+            return score + A_SCORE_GAP;
+        }
+        return score;
+    }
+
     public boolean cannotBust() {
-        return score.cannotBust();
+        return calculateScore() <= BUST_THRESHOLD;
+    }
+
+    public boolean cannotBust(int score) {
+        return score <= BUST_THRESHOLD;
     }
 
     public boolean cannotDealerHit() {
-        return score.cannotDealerHit();
+        return calculateScore() <= DEALER_HIT_THRESHOLD;
     }
 
-    public boolean hasScore(int comparedScore) {
-        return score.hasScore(comparedScore);
+    public boolean hasScore(int score) {
+        return calculateScore() == score;
     }
 
     public Card pickFirstCard() {
