+++ src/main/java/domain/result/Judge.java
@@ -2,7 +2,6 @@ package domain.result;
 
 import domain.Profit;
 import domain.WinState;
-import domain.cards.Score;
 import domain.gamer.Dealer;
 import domain.gamer.Gamers;
 import domain.gamer.Player;
@@ -52,13 +51,12 @@ public class Judge {
     }
 
     private WinState decidePlayerWinState(Player player, Dealer dealer) {
-        Score playerScore = player.finalizeCardsScore();
-        Score dealerScore = dealer.finalizeCardsScore();
-        int gapOfScore = playerScore.compareScore(dealerScore);
-        if (gapOfScore > 0) {
+        int playerScore = player.finalizeCardsScore();
+        int dealerScore = dealer.finalizeCardsScore();
+        if (playerScore > dealerScore) {
             return WinState.WIN;
         }
-        if (gapOfScore < 0) {
+        if (playerScore < dealerScore) {
             return WinState.LOSE;
         }
         return WinState.DRAW;
